pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "myl-discovery"
authors = [
  {name = "Philipp Schmitt", email = "philipp@schmitt.co"},
]
description = "email autodiscovery"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["email-discovery", "autodiscover", "autoconfig",  "smtp", "imap", "email", "discovery"]
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python :: 3",
]
dependencies = [
  "dnspython",
  "requests",
  "xmltodict"
]
version = "0.1.0"

[tool.black]
line-length = 79
